Cleaning and moving props part 17
